+++ src/documentmanager.cpp
@@ -1,6 +1,6 @@
 /***********************************************************************
  *
- * Copyright (C) 2014-2020 wereturtle
+ * Copyright (C) 2014-2021 wereturtle
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -28,10 +28,12 @@
 #include <QMessageBox>
 #include <QPair>
 #include <QString>
+#include <QStandardPaths>
 #include <QtConcurrentRun>
 #include <QTextDocument>
 #include <QTextStream>
 #include <QTimer>
+#include <QDebug>
 
 #include "documenthistory.h"
 #include "documentmanager.h"
@@ -64,6 +66,10 @@ public:
     {
 
     }
+
+    const QString draftName = QObject::tr("untitled");
+
+    QString draftLocation;
 
     DocumentManager *q_ptr;
     MarkdownDocument *document;
@@ -99,8 +105,14 @@ public:
     */
     void saveFile();
 
+    /*
+    * Handles any errors or tidying up after an asynchronous save operation.
+    */
     void onSaveCompleted();
 
+    /*
+    * Handles the event where a file as been modified externally on disk.
+    */
     void onFileChangedExternally(const QString &path);
 
     /*
@@ -150,7 +162,22 @@ public:
     */
     void backupFile(const QString &filePath) const;
 
+    /*
+    * Handles autosave operation upon autosave timer expiration.
+    */
     void autoSaveFile();
+
+    /*
+    * Returns true if document is named as "untitled" and located in the
+    * configured draft location.
+    */
+    bool documentIsDraft();
+
+    /*
+    * Creates a draft (named "untitled-##.md") document in the configured
+    * draft location.
+    */
+    void createDraft();
 };
 
 const QString DocumentManagerPrivate::FILE_CHOOSER_FILTER =
@@ -176,6 +203,9 @@ DocumentManager::DocumentManager
     d->documentModifiedNotifVisible = false;
     d->saveFutureWatcher = new QFutureWatcher<QString>(this);
 
+    d->draftLocation =
+        QStandardPaths::writableLocation(QStandardPaths::DocumentsLocation);
+
     d->fileWatcher = new QFileSystemWatcher(this);
     d->document = (MarkdownDocument *) editor->document();
 
@@ -183,44 +213,38 @@ DocumentManager::DocumentManager
     d->autoSaveTimer = new QTimer(this);
     d->autoSaveTimer->start(60000);
 
-    this->connect
-    (
-        d->autoSaveTimer,
+    this->connect(d->autoSaveTimer,
         &QTimer::timeout,
         [d]() {
             d->autoSaveFile();
         }
     );
 
-
-    connect
-    (
-        d->document,
+    connect(d->document,
         &MarkdownDocument::modificationChanged,
         [this, d](bool modified) {
-            if
-            (
-                d->document->isNew()
-                || d->document->isReadOnly()
-                || !d->autoSaveEnabled
-            ) {
+            if (d->document->isReadOnly()
+                    || !d->autoSaveEnabled) {
                 emit documentModifiedChanged(modified);
+            }
+
+            if (modified
+                    && d->autoSaveEnabled 
+                    && d->document->isNew()
+                    && (d->document->characterCount() > 0)) {
+                d->createDraft();
             }
         }
     );
 
-    this->connect
-    (
-        d->saveFutureWatcher,
+    this->connect(d->saveFutureWatcher,
         &QFutureWatcher<QString>::finished,
         [d]() {
             d->onSaveCompleted();
         }
     );
 
-    this->connect
-    (
-        d->fileWatcher,
+    this->connect(d->fileWatcher,
         &QFileSystemWatcher::fileChanged,
         [d](const QString & path) {
             d->onFileChangedExternally(path);
@@ -249,20 +273,6 @@ bool DocumentManager::autoSaveEnabled() const
     return d->autoSaveEnabled;
 }
 
-bool DocumentManager::fileBackupEnabled() const
-{
-    Q_D(const DocumentManager);
-    
-    return d->createBackupOnSave;
-}
-
-void DocumentManager::setFileHistoryEnabled(bool enabled)
-{
-    Q_D(DocumentManager);
-    
-    d->fileHistoryEnabled = enabled;
-}
-
 void DocumentManager::setAutoSaveEnabled(bool enabled)
 {
     Q_D(DocumentManager);
@@ -270,10 +280,23 @@ void DocumentManager::setAutoSaveEnabled(bool enabled)
     d->autoSaveEnabled = enabled;
 
     if (enabled) {
+        if (d->document->isNew()
+            && (d->document->characterCount() > 0)
+            && d->document->isModified()) {
+            d->createDraft();
+        }
+
         emit documentModifiedChanged(false);
     } else if (d->document->isModified()) {
         d->document->setModified(false);
     }
+}
+
+bool DocumentManager::fileBackupEnabled() const
+{
+    Q_D(const DocumentManager);
+    
+    return d->createBackupOnSave;
 }
 
 void DocumentManager::setFileBackupEnabled(bool enabled)
@@ -281,6 +304,26 @@ void DocumentManager::setFileBackupEnabled(bool enabled)
     Q_D(DocumentManager);
     
     d->createBackupOnSave = enabled;
+}
+
+void DocumentManager::setDraftLocation(const QString &directory) 
+{
+    Q_D(DocumentManager);
+
+    QDir draftDir(directory);
+
+    if (!draftDir.exists()) {
+        draftDir.mkpath(draftDir.path());
+    }
+
+    d->draftLocation = draftDir.absolutePath();
+}
+
+void DocumentManager::setFileHistoryEnabled(bool enabled)
+{
+    Q_D(DocumentManager);
+    
+    d->fileHistoryEnabled = enabled;
 }
 
 void DocumentManager::open(const QString &filePath)
@@ -506,14 +549,17 @@ bool DocumentManager::close()
         cursor.setPosition(0);
         d->editor->setTextCursor(cursor);
 
+        d->document->blockSignals(true);
         d->document->setPlainText("");
+        d->document->blockSignals(false);
         d->document->clearUndoRedoStacks();
         d->editor->setReadOnly(false);
         d->document->setReadOnly(false);
         d->setFilePath(QString());
         d->document->setModified(false);
 
-        if (d->fileHistoryEnabled && !documentIsNew) {
+        if (d->fileHistoryEnabled && 
+                (!documentIsNew || this->autoSaveEnabled())) {
             DocumentHistory history;
             history.add
             (
@@ -620,21 +666,6 @@ void DocumentManagerPrivate::onFileChangedExternally(const QString &path)
     }
 }
 
-void DocumentManagerPrivate::autoSaveFile()
-{
-    Q_Q(DocumentManager);
-
-    if
-    (
-        this->autoSaveEnabled &&
-        !this->document->isNew() &&
-        !this->document->isReadOnly() &&
-        this->document->isModified()
-    ) {
-        q->save();
-    }
-}
-
 void DocumentManagerPrivate::saveFile()
 {
     Q_Q(DocumentManager);
@@ -679,9 +710,7 @@ bool DocumentManagerPrivate::loadFile(const QString &filePath)
     QFile inputFile(filePath);
 
     if (!inputFile.open(QIODevice::ReadOnly)) {
-        MessageBoxHelper::critical
-        (
-            editor,
+        MessageBoxHelper::critical(editor,
             QObject::tr("Could not read %1").arg(filePath),
             inputFile.errorString()
         );
@@ -700,7 +729,9 @@ bool DocumentManagerPrivate::loadFile(const QString &filePath)
 
     document->clearUndoRedoStacks();
     document->setUndoRedoEnabled(false);
+    document->blockSignals(true);
     document->setPlainText("");
+    document->blockSignals(false);
 
     QApplication::setOverrideCursor(Qt::WaitCursor);
     emit q->operationStarted(QObject::tr("opening %1").arg(filePath));
@@ -715,16 +746,8 @@ bool DocumentManagerPrivate::loadFile(const QString &filePath)
 
     QString text = inStream.readAll();
 
-    editor->setPlainText(text);
-    editor->navigateDocument(0);
-    emit q->operationUpdate();
-
-    document->setUndoRedoEnabled(true);
-
     if (QFile::NoError != inputFile.error()) {
-        MessageBoxHelper::critical
-        (
-            editor,
+        MessageBoxHelper::critical(editor,
             QObject::tr("Could not read %1").arg(filePath),
             inputFile.errorString()
         );
@@ -732,6 +755,13 @@ bool DocumentManagerPrivate::loadFile(const QString &filePath)
         inputFile.close();
         return false;
     }
+    
+    setFilePath(filePath);
+    editor->setPlainText(text);
+    editor->navigateDocument(0);
+    emit q->operationUpdate();
+
+    document->setUndoRedoEnabled(true);
 
     inputFile.close();
 
@@ -742,7 +772,6 @@ bool DocumentManagerPrivate::loadFile(const QString &filePath)
         editor->navigateDocument(0);
     }
 
-    setFilePath(filePath);
     editor->setReadOnly(false);
 
     if (!fileInfo.isWritable()) {
@@ -949,4 +978,49 @@ void DocumentManagerPrivate::backupFile(const QString &filePath) const
                   file.errorString().toLatin1().data());
     }
 }
+
+void DocumentManagerPrivate::autoSaveFile()
+{
+    Q_Q(DocumentManager);
+
+    if
+    (
+        this->autoSaveEnabled &&
+        !this->document->isNew() &&
+        !this->document->isReadOnly() &&
+        this->document->isModified()
+    ) {
+        q->save();
+    }
+}
+
+bool DocumentManagerPrivate::documentIsDraft()
+{
+    if (document->isNew()) {
+        return false;
+    }
+
+    QFileInfo info(document->filePath());
+
+    return ((info.dir().absolutePath() == draftLocation)
+        && (info.baseName().startsWith(draftName)));
+}
+
+void DocumentManagerPrivate::createDraft()
+{
+    if (document->isNew()) {
+        int i = 1;
+        QString draftPath;
+
+        // Make sure draft file name is unique.
+        do {
+            draftPath = draftLocation + "/" + draftName + "-" + QString::number(i) + ".md";
+            i++;
+        } while (QFileInfo(draftPath).exists());
+
+        this->setFilePath(draftPath);
+        saveFile();
+    }
+}
+
 }