+++ src/preferencesdialog.cpp
@@ -1,6 +1,6 @@
 /***********************************************************************
  *
- * Copyright (C) 2016-2020 wereturtle
+ * Copyright (C) 2016-2021 wereturtle
  * Copyright (C) 2009, 2010, 2011, 2012, 2013, 2014 Graeme Gott <graeme@gottcode.org>
  *
  * This program is free software: you can redistribute it and/or modify
@@ -20,6 +20,7 @@
 
 #include <QCheckBox>
 #include <QComboBox>
+#include <QDesktopServices>
 #include <QDialogButtonBox>
 #include <QFormLayout>
 #include <QGroupBox>
@@ -27,6 +28,7 @@
 #include <QPushButton>
 #include <QSpinBox>
 #include <QTabWidget>
+#include <QUrl>
 #include <QVBoxLayout>
 
 #include "appsettings.h"
@@ -240,6 +242,16 @@ QWidget *PreferencesDialogPrivate::initializeGeneralTab()
     backupCheckBox->setChecked(appSettings->backupFileEnabled());
     connect(backupCheckBox, SIGNAL(toggled(bool)), appSettings, SLOT(setBackupFileEnabled(bool)));
     savingGroupLayout->addRow(backupCheckBox);
+
+    QPushButton *openDraftDirButton = new QPushButton(tr("View saved drafts..."));
+    q->connect(
+        openDraftDirButton,
+        &QPushButton::clicked,
+        [this]() {
+            QDesktopServices::openUrl(QUrl(appSettings->draftLocation()));
+        }
+    );
+    savingGroupLayout->addRow(openDraftDirButton);
 
     QGroupBox *historyGroupBox = new QGroupBox(tr("History"));
     tabLayout->addWidget(historyGroupBox);