+++ src/themerepository.cpp
@@ -1,6 +1,6 @@
 /***********************************************************************
  *
- * Copyright (C) 2014-2020 wereturtle
+ * Copyright (C) 2014-2021 wereturtle
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -27,6 +27,7 @@
 #include <QJsonObject>
 #include <QJsonValue>
 #include <QMetaEnum>
+#include <QMultiMap>
 #include <QSettings>
 #include <QStack>
 #include <QString>
@@ -97,11 +98,11 @@ ThemeRepository::ThemeRepository()
     d->themeDirectory.setNameFilters(QStringList("*.json"));
 
     QFileInfoList themeFiles = d->themeDirectory.entryInfoList();
-    QMap<QString, QString> sortedThemes;
+    QMultiMap<QString, QString> sortedThemes;
 
     for (int i = 0; i < themeFiles.size(); i++) {
         QString baseName = themeFiles[i].baseName();
-        sortedThemes.insertMulti(baseName.toLower(), baseName);
+        sortedThemes.insert(baseName.toLower(), baseName);
     }
 
     d->customThemeNames = sortedThemes.values();