+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* Untitled documents are now autosaved to a draft folder when autosave is enabled.
+* Added preferences button to open the draft folder location where untitled documents are autosaved.
+* Updated Brazilian Portuguese translation.
+
 ## [2.0.2] - 2021-06-27
 
 ### Fixed